Aggresive make buffering on two channels + compile -> server render + chrome ws -> client render
