+++ addons/frontend/package.json
@@ -13,7 +13,8 @@
   },
   "dependencies": {
     "@myriaddreamin/typst.ts": "0.4.0-rc7",
-    "@myriaddreamin/typst-ts-renderer": "0.4.0-rc7"
+    "@myriaddreamin/typst-ts-renderer": "0.4.0-rc7",
+    "rxjs": "^7.8.1"
   },
   "devDependencies": {
     "typescript": "^5.0.2",
